requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Redis
redis[hiredis]==5.2.1

# Auth
pyjwt==2.10.1
passlib[argon2]==1.7.4
argon2-cffi==23.1.0

# Encryption
cryptography==44.0.0

# Validation / utilities
pydantic==2.10.3
pydantic-settings==2.7.0
python-multipart==0.0.19

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
